Defer the properties validation to the next usage of the newConnection method, in order to allow the properties to be set in whatever order
Using a state Object that contains the configuration set by the user.
Uppon first use of the newConnection method, the state is copied and validated.
When a property setter is called, the validated state is set to null.
Uppon next use of the newConnection method, if the validated State is already build, validation is skipped, otherwise it is rebuild and validated.